+++ common/constants/shared.ts
