+++ common/utils/query_utils.ts
@@ -9,10 +9,10 @@
  * GitHub history for details.
  */
 
-export const getIndexPatternFromRawQuery = (query: string) => {
+export const getIndexPatternFromRawQuery = (query: string) : string => {
   const indexMatches = query.match('\=(.*)');
   if (indexMatches) {
-    return indexMatches[0]?.slice(1).split('|')[0]?.trim();
+    return indexMatches[0]?.slice(1)?.split('|')[0]?.trim();
   }
   return '';
 };
