+++ public/plugin.ts
@@ -24,7 +24,7 @@ import {
   observabilityID,
   observabilityTitle,
   observabilityPluginOrder
-} from '../common/index';
+} from '../common/constants/shared';
 import PPLService from './services/requests/ppl';
 
 export class ObservabilityPlugin implements Plugin<ObservabilitySetup, ObservabilityStart> {
